Sesión de usuario, Server y Client side validation

+++ src/app/api/todos/route.ts
@@ -1,4 +1,5 @@
 
+import { getUserSessionServer } from "@/auth/actions/auth-acctions";
 import prisma from "@/lib/prisma";
 import { NextResponse, NextRequest } from "next/server";
 import * as yup from 'yup'
@@ -36,11 +37,18 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) { 
 
+  const user = await getUserSessionServer()
+
+  if ( !user ){
+    return NextResponse.json('No utorizado',{status:401})
+  }
+
+
   try {
     // Parsea el cuerpo de la solicitud como JSON y valida que cumpla con el esquema definido
     const { complete, description } = await postSchema.validate (await request.json())
 
-    const todo = await prisma.todo.create({ data: {complete, description} })    
+    const todo = await prisma.todo.create({ data: {complete, description, userId: user.id} })    
         
     return NextResponse.json(todo)
     
@@ -55,9 +63,15 @@ export async function POST(request: Request) {
 
 export async function DELETE(request: Request) { 
 
+  const user = await getUserSessionServer()
+
+  if ( !user ){
+    return NextResponse.json('No utorizado',{status:401})
+  }
+
   try {
 
-   await prisma.todo.deleteMany({ where: {complete: true} })    
+   await prisma.todo.deleteMany({ where: {complete: true, userId:user.id} })    
         
     return NextResponse.json('Borrados')
